This PR add support for reading the elevation mapping mode (LatestPass or MinimumElevation) and storing that as a set of time based events in the TRex project.

It also contains significant unit test framework development to support unit tests of underlying query functionality against data stored in a TRex project that may be constructed within the unit test in two ways:

1. By import of one or more TAG files into an isolated project context that has no Ignite or other external dependencies but which preserves cluster compute layer capabilities for sub grid request processing.
2. By direct/manual creation of a project from the ground up with a precise set of data tailored to support unit testing of business logic in question by exercising it with sub grid requests in the same manner as 1 above.

Both approaches are demonstrated in the elevation mapping mode unit tests.

Related work items: #78096, #79096

+++ src/service/TRex/src/netstandard/VSS.TRex.TAGFiles/Classes/ValueMatcher/CoordinateSystem/TAGCoordinateSystemTypeValueMatcher.cs
@@ -25,7 +25,7 @@ namespace VSS.TRex.TAGFiles.Classes.ValueMatcher.CoordinateSystem
                 return false;
             }
 
-            if (!Enum.IsDefined(typeof(CoordinateSystemType), (int)value))
+            if (!Enum.IsDefined(typeof(CoordinateSystemType), (byte)value))
             {
                 return false;
             }
